refactor(storage_impl): Move Conversion impl for business_profile to storage_impl crate
